hiring_system/requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.20
crewai>=0.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# API
fastapi>=0.108.0
uvicorn>=0.24.0
starlette>=0.31.1
pydantic>=2.0.0

# API clients
openai>=1.3.0  # For GPT-4o API integration

# AWS integration
boto3>=1.28.0
botocore>=1.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Graph processing
networkx>=3.0

# Utilities
matplotlib>=3.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# For MPS (Metal Performance Shaders) support on Apple Silicon
torch>=2.0.0
